refactor(deploy): git mv deploy/ and the shared version spine into the new crates
Pure rename commit — no content edits. Rewiring is the next commit.

deploy/            -> crates/homeboy-deploy/src/   (mod.rs -> lib.rs)
release/version.rs -> crates/homeboy-version/src/
release/version/   -> crates/homeboy-version/src/version/
release/changelog/ -> crates/homeboy-version/src/changelog/
release/scope.rs   -> crates/homeboy-version/src/scope.rs

version/guard.rs stays in homeboy-release (as version_guard.rs): it is the
only piece of the version module that reaches back into release-only code
(planning_worktree::derived_release_lockfiles), and deploy never calls it.

Refs #11144
