+++ crates/homeboy-deploy/src/content_manifest.rs
