+++ crates/homeboy-deploy/src/generated_artifacts.rs
