+++ crates/homeboy-deploy/src/path_roots.rs
