+++ crates/homeboy-version/src/scope.rs
