+++ crates/homeboy-version/src/version/default_pattern_for_file.rs
